+++ docs/search/all_13.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['vector2_433',['Vector2',['../classraylib_1_1_vector2.html',1,'raylib']]],
-  ['vector3_434',['Vector3',['../classraylib_1_1_vector3.html',1,'raylib']]],
-  ['vector4_435',['Vector4',['../classraylib_1_1_vector4.html',1,'raylib']]],
-  ['vrstereoconfig_436',['VrStereoConfig',['../classraylib_1_1_vr_stereo_config.html',1,'raylib']]]
+  ['vector2_437',['Vector2',['../classraylib_1_1_vector2.html',1,'raylib']]],
+  ['vector3_438',['Vector3',['../classraylib_1_1_vector3.html',1,'raylib']]],
+  ['vector4_439',['Vector4',['../classraylib_1_1_vector4.html',1,'raylib']]],
+  ['vrstereoconfig_440',['VrStereoConfig',['../classraylib_1_1_vr_stereo_config.html',1,'raylib']]]
 ];