+++ docs/search/classes_8.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['ray_463',['Ray',['../classraylib_1_1_ray.html',1,'raylib']]],
-  ['raycollision_464',['RayCollision',['../classraylib_1_1_ray_collision.html',1,'raylib']]],
-  ['raylibexception_465',['RaylibException',['../classraylib_1_1_raylib_exception.html',1,'raylib']]],
-  ['rectangle_466',['Rectangle',['../classraylib_1_1_rectangle.html',1,'raylib']]],
-  ['rendertexture_467',['RenderTexture',['../classraylib_1_1_render_texture.html',1,'raylib']]]
+  ['ray_467',['Ray',['../classraylib_1_1_ray.html',1,'raylib']]],
+  ['raycollision_468',['RayCollision',['../classraylib_1_1_ray_collision.html',1,'raylib']]],
+  ['raylibexception_469',['RaylibException',['../classraylib_1_1_raylib_exception.html',1,'raylib']]],
+  ['rectangle_470',['Rectangle',['../classraylib_1_1_rectangle.html',1,'raylib']]],
+  ['rendertexture_471',['RenderTexture',['../classraylib_1_1_render_texture.html',1,'raylib']]]
 ];