+++ docs/search/classes_b.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['vector2_471',['Vector2',['../classraylib_1_1_vector2.html',1,'raylib']]],
-  ['vector3_472',['Vector3',['../classraylib_1_1_vector3.html',1,'raylib']]],
-  ['vector4_473',['Vector4',['../classraylib_1_1_vector4.html',1,'raylib']]],
-  ['vrstereoconfig_474',['VrStereoConfig',['../classraylib_1_1_vr_stereo_config.html',1,'raylib']]]
+  ['vector2_475',['Vector2',['../classraylib_1_1_vector2.html',1,'raylib']]],
+  ['vector3_476',['Vector3',['../classraylib_1_1_vector3.html',1,'raylib']]],
+  ['vector4_477',['Vector4',['../classraylib_1_1_vector4.html',1,'raylib']]],
+  ['vrstereoconfig_478',['VrStereoConfig',['../classraylib_1_1_vr_stereo_config.html',1,'raylib']]]
 ];