+++ docs/search/classes_c.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['wave_475',['Wave',['../classraylib_1_1_wave.html',1,'raylib']]],
-  ['window_476',['Window',['../classraylib_1_1_window.html',1,'raylib']]]
+  ['wave_479',['Wave',['../classraylib_1_1_wave.html',1,'raylib']]],
+  ['window_480',['Window',['../classraylib_1_1_window.html',1,'raylib']]]
 ];