+++ docs/search/functions_15.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['_7eaudiodevice_901',['~AudioDevice',['../classraylib_1_1_audio_device.html#aab60bade54ebe2fc41e567d0023047d9',1,'raylib::AudioDevice']]],
-  ['_7emusic_902',['~Music',['../classraylib_1_1_music.html#a6fb0e1cb0807c33e952bdd8c5028fa16',1,'raylib::Music']]],
-  ['_7evrstereoconfig_903',['~VrStereoConfig',['../classraylib_1_1_vr_stereo_config.html#affd207a5267f0ea9c48d92dcfd72edea',1,'raylib::VrStereoConfig']]],
-  ['_7ewave_904',['~Wave',['../classraylib_1_1_wave.html#a545a0afb559e87f42cdedcda263452ba',1,'raylib::Wave']]],
-  ['_7ewindow_905',['~Window',['../classraylib_1_1_window.html#a6071f03b18e0f2d3817b0da3699f24af',1,'raylib::Window']]]
+  ['_7eaudiodevice_911',['~AudioDevice',['../classraylib_1_1_audio_device.html#aab60bade54ebe2fc41e567d0023047d9',1,'raylib::AudioDevice']]],
+  ['_7emusic_912',['~Music',['../classraylib_1_1_music.html#a6fb0e1cb0807c33e952bdd8c5028fa16',1,'raylib::Music']]],
+  ['_7evrstereoconfig_913',['~VrStereoConfig',['../classraylib_1_1_vr_stereo_config.html#affd207a5267f0ea9c48d92dcfd72edea',1,'raylib::VrStereoConfig']]],
+  ['_7ewave_914',['~Wave',['../classraylib_1_1_wave.html#a545a0afb559e87f42cdedcda263452ba',1,'raylib::Wave']]],
+  ['_7ewindow_915',['~Window',['../classraylib_1_1_window.html#a6071f03b18e0f2d3817b0da3699f24af',1,'raylib::Window']]]
 ];